refactor: move `hasFocus` logic to `FocusableDwcComponent`
