+++ dwcj-engine/src/test/java/org/dwcj/component/button/AbstractDwcButtonTest.java
@@ -42,24 +42,6 @@ class AbstractDwcButtonTest {
   }
 
   @Nested
-  @DisplayName("Focus API")
-  class FocusApi {
-    @Test
-    @DisplayName("hasFocus when control is defined")
-    void hasFocusWhenControlIsDefined() throws BBjException {
-      doReturn("true").when(control).getClientProperty("hasFocus");
-      assertTrue(component.hasFocus());
-    }
-
-    @Test
-    @DisplayName("hasFocus when control is null")
-    void hasFocusWhenControlIsNull() throws BBjException, IllegalAccessException {
-      ReflectionUtils.nullifyControl(component);
-      assertFalse(component.hasFocus());
-    }
-  }
-
-  @Nested
   @DisplayName("DisableOnclick API")
   class DisableOnclickApi {
     @Test
