+++ dwcj-engine/src/test/java/org/dwcj/component/optioninput/AbstractDwcOptionInputTest.java
@@ -115,24 +115,6 @@ class AbstractDwcOptionInputTest {
   }
 
   @Nested
-  @DisplayName("Focus API")
-  class FocusApi {
-    @Test
-    @DisplayName("hasFocus when control is defined")
-    void hasFocusWhenControlIsDefined() throws BBjException {
-      doReturn("true").when(control).getClientProperty("hasFocus");
-      assertTrue(component.hasFocus());
-    }
-
-    @Test
-    @DisplayName("hasFocus when control is null")
-    void hasFocusWhenControlIsNull() throws BBjException, IllegalAccessException {
-      ReflectionUtils.nullifyControl(component);
-      assertFalse(component.hasFocus());
-    }
-  }
-
-  @Nested
   @DisplayName("onAttach behavior")
   class OnAttach {
